Return to automatic drop
We now issue the `DeleteSession` command when the client is dropped, and
return an addition future to the user from `Client::new` that they
should wait for after everything else is done. This second future will
only resolve once all futures spawned off of that `Client` have
resolved, *and* the `Client` has been dropped, *and* the `DeleteSession`
command has been issued.