Move to tokio and serde_json
Phew, this is quite the change. This moves everything from tokio-core
and rustc-serialize to tokio and serde_json. With that comes a number of
API changes. The biggest among them is that `Client` is now thread-safe
and `Clone`, but all receives now take `&mut self` or `self` (for
navigation and waiting). This should make transitioning to async/await
better, but in the meantime it will require some code juggling for
users. Sorry in advance for that.

The biggest internal change is that all the WebDriver protocol work is
now driven by a new `Session` thingy that is spawned onto the tokio
runtime when the user tries to create a fresh `Client`. It implements
Future, and just receives and processes one request at a time to
completion. Requests are sent over a sync::mpsc, which is why `Client`
can now be thread-safe and cloneable. This also tidies up cleanup a lot,
since the Session will make sure to tidy itself up whenever the are no
`Client`s with `mpsc::Sender` handles left.

Fixes #28
Fixes #20
Closes #34