Make chromedriver work on CI again
chromedriver was defaulting to Google Chrome, which is apparently
installed by default on Travis. When chromedriver was recently bumped to
v74 (and Chrome was not), it would error out with a message along the
lines of

  This version of ChromeDriver only supports Chrome version 74

This patch fixes that issue by explicitly pointing fantoccini at
/usr/bin/chromium{-browser}. Since we install both chromedriver and
chromium from the chromium-* packages, they _should_ always be the same
version.