Bump webdriver and remove server feature
`webdriver 0.43` made the server aspects of the code optional
(https://bugzilla.mozilla.org/show_bug.cgi?id=1684226). We take
advantage of that to get rid of `webdriver`'s `warp` dependency which
brought in an old version of tokio.

This change is unfortunately backwards-incompatible because we publicly
expose types from the `webdriver` crate.