Simplify new wait interface
Specifically, this removes the generic trait-based interface, because we
only want to provide wait methods for finders that we _may_ be able to
optimize the behavior of in the future. For example, `for_element` may
be able to leverage JavaScript observers to avoid the spin loop in favor
of an event-driven waiting pattern. Same thing with `for_url`. If users
just want to retry an operation, they can do so in their own code.