Prepare for 0.22 release (#294)
Updates all dependencies to latest major version. This includes the
webdriver crate, which now generates floating point numbers for pointer
move operations. These are in turn rejected by older versions of Chrome
and Firefox. See https://bugzilla.mozilla.org/show_bug.cgi?id=1959464
for details. As such, this is breaking in the sense that older versions
of said browsers are no longer supported.

This also bumps the crate's MSRV, but only due to dependencies, not code
changes.

Note that minimal-versions in CI will remain broken until webdriver
upgrades to a newer version of icu_segmenter (specifically: 2.0) so that
they move on from the zerovec version that has this stanza:
<https://github.com/unicode-org/icu4x/blob/icu%401.5.0/utils/zerovec/Cargo.toml#L33-L38>